+++ index.ts
@@ -1,0 +1,3 @@
+import UpdateSlackStatus from './slack-operations/update-status';
+
+UpdateSlackStatus();
