+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon index.js"
+    "start": "nodemon index.ts"
   },
   "keywords": [],
   "author": "",
@@ -17,7 +17,10 @@
     "moment": "^2.29.3"
   },
   "devDependencies": {
+    "@types/nodemon": "^1.19.1",
     "dotenv": "^16.0.1",
-    "nodemon": "^2.0.16"
+    "nodemon": "^2.0.16",
+    "ts-node": "^10.8.1",
+    "typesync": "^0.9.1"
   }
 }
